feat(api): write the authenticationComplement in DB at every authentication
This way the authenticationComplement is always synchronized with the current
state of the user, especially when there is no more claimsToStore configured.
But there is a useless overload when the authenticationComplement has not
changed since the last time. A performance improvement would be to write the
authenticationMethod in DB only if the authenticationComplement has changed, but
this will be done in another PR if performance problems arise.